Updated reqs, working extracts prompt generation
