+++ environment.yml
@@ -1,0 +1,99 @@
+name: sba
+channels:
+  - pytorch
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - _openmp_mutex=5.1=1_gnu
+  - blas=1.0=mkl
+  - bzip2=1.0.8=h7b6447c_0
+  - ca-certificates=2022.07.19=h06a4308_0
+  - charset-normalizer=2.0.4=pyhd3eb1b0_0
+  - cudatoolkit=11.3.1=h2bc3f7f_2
+  - ffmpeg=4.3=hf484d3e_0
+  - freetype=2.11.0=h70c0345_0
+  - giflib=5.2.1=h7b6447c_0
+  - gmp=6.2.1=h295c915_3
+  - gnutls=3.6.15=he1e5248_0
+  - idna=3.3=pyhd3eb1b0_0
+  - intel-openmp=2021.4.0=h06a4308_3561
+  - jpeg=9e=h7f8727e_0
+  - lame=3.100=h7b6447c_0
+  - lcms2=2.12=h3be6417_0
+  - ld_impl_linux-64=2.38=h1181459_1
+  - lerc=3.0=h295c915_0
+  - libdeflate=1.8=h7f8727e_5
+  - libffi=3.3=he6710b0_2
+  - libgcc-ng=11.2.0=h1234567_1
+  - libgomp=11.2.0=h1234567_1
+  - libiconv=1.16=h7f8727e_2
+  - libidn2=2.3.2=h7f8727e_0
+  - libpng=1.6.37=hbc83047_0
+  - libstdcxx-ng=11.2.0=h1234567_1
+  - libtasn1=4.16.0=h27cfd23_0
+  - libtiff=4.4.0=hecacb30_0
+  - libunistring=0.9.10=h27cfd23_0
+  - libuuid=1.0.3=h7f8727e_2
+  - libwebp=1.2.2=h55f646e_0
+  - libwebp-base=1.2.2=h7f8727e_0
+  - lz4-c=1.9.3=h295c915_1
+  - mkl=2021.4.0=h06a4308_640
+  - mkl_fft=1.3.1=py310hd6ae3a3_0
+  - mkl_random=1.2.2=py310h00e6091_0
+  - ncurses=6.3=h5eee18b_3
+  - nettle=3.7.3=hbbd107a_1
+  - numpy-base=1.23.1=py310hcba007f_0
+  - openh264=2.1.1=h4ff587b_0
+  - openssl=1.1.1q=h7f8727e_0
+  - pycparser=2.21=pyhd3eb1b0_0
+  - pyopenssl=22.0.0=pyhd3eb1b0_0
+  - python=3.10.4=h12debd9_0
+  - pytorch=1.12.1=py3.10_cuda11.3_cudnn8.3.2_0
+  - pytorch-mutex=1.0=cuda
+  - readline=8.1.2=h7f8727e_1
+  - six=1.16.0=pyhd3eb1b0_1
+  - sqlite=3.39.2=h5082296_0
+  - tk=8.6.12=h1ccaba5_0
+  - typing_extensions=4.3.0=py310h06a4308_0
+  - tzdata=2022a=hda174b7_0
+  - wheel=0.37.1=pyhd3eb1b0_0
+  - xz=5.2.5=h7f8727e_1
+  - zlib=1.2.12=h7f8727e_2
+  - zstd=1.5.2=ha4553b6_0
+  - pip:
+    - brotlipy==0.7.0
+    - certifi==2022.6.15
+    - cffi==1.15.1
+    - click==8.1.3
+    - cryptography==37.0.1
+    - diffusers==0.3.0
+    - filelock==3.8.0
+    - huggingface-hub==0.9.1
+    - importlib-metadata==4.12.0
+    - joblib==1.1.0
+    - lxml==4.9.1
+    - mkl-fft==1.3.1
+    - mkl-random==1.2.2
+    - mkl-service==2.4.0
+    - nltk==3.7
+    - numpy==1.23.1
+    - packaging==21.3
+    - pillow==9.2.0
+    - pip==22.1.2
+    - pyparsing==3.0.9
+    - pysocks==1.7.1
+    - python-docx==0.8.11
+    - pyyaml==6.0
+    - regex==2022.8.17
+    - requests==2.28.1
+    - setuptools==63.4.1
+    - tokenizers==0.12.1
+    - torch==1.12.1
+    - torchaudio==0.12.1
+    - torchvision==0.13.1
+    - tqdm==4.64.1
+    - transformers==4.21.3
+    - typing-extensions==4.3.0
+    - urllib3==1.26.11
+    - zipp==3.8.1
+prefix: /cvgl2/u/andreyk/miniconda3/envs/sba
