+++ inputs/extracts_prompt_prefix.txt
@@ -1,0 +1,23 @@
+Suggest descriptions of images to illustrate for the following paragraphs from short stories:
+~
+You notice things others don’t when you spend your entire life stuck in a tower. Rapunzel knew the best combination of hair products from all available in the realm. She knew the health of the crops by the color of the sunlit fields. She thought it curious that across the Queen’s valley, a stone tower stood apart from the rest of the village, similar to the one in which she was imprisoned. She used her hand mirror to flash light at the other tower.---rapunzel in a tower, using her hand mirror to flash light at the other tower. Watercolor
+~
+The first earthquake was strong enough to wake Zeke early in the morning. His house was unharmed, but the garden shed in his backyard collapsed into rubble. After making coffee while listening to news reports about the earthquake, Zeke took his mug out back and surveyed the damage. Zeke found a small, blue lamp in the debris, and he was curious to discover what it was.---oil painting of an earthquake: the garden shed in his backyard collapsed into rubble. Zeke found a small, blue lamp in the debris
+~
+“My second wish is that you help me expand the Earth.” Zeke was excited about making the wish, but he was a little concerned about the magnitude of his request. “The Earth is getting larger,” the genie confirmed. “But you can’t see it yet.” “When will it spread?” Zeke asked. “How long will it take?” The genie tilted his head and said, “The Earth will be double its current size in ten months, twenty-three days, and eight hours.”---impressionist painting of the earth getting larger and a ring around the earth
+~
+"Hey Tom" said Jerry. "What's on your mind Jerry" said Tom. "It's like this" said Jerry. "We've been in these cartoons since the early 1940s, and frankly I'm sick of it." "That's nothing new" said Tom. "After all we've been torturing each other for over 80 years, and it seems like it never ends." "Well, I've got an idea" said Jerry. "I've been reading this book called Myth of Sisyphus, and I think we should roll a giant boulder up the hill and keep pushing it back down again."---Tom and Jerry discussing the Myth of Sisyphus, detailed pencil sketch, dark 
+~
+I was in the lab the day that the greatest discovery in the history of science was made (though I acknowledge that I may be biased). Our lab put multiverse theory into practice to create controlled, self-contained micro-universes within our own. We programmed the physics inside and watched planets evolve over massive geologic time scales. ---a crystal ball containing a micro-universe, digital 3D art, high resolution, vivid
+~
+Once, there lived a man who hated mowing his lawn, so he bought a small herd of goats to do it for him. One of the goats was cleverer than the rest. It learned how to ask the man for a drink whenever it was thirsty. The man would bring the goat a drink, and the goat would mow the lawn.---A wealthy man bringing a goat a drink so the goat would mow the lawn, oils painting, renaissance, fable
+~
+It was all premature, though. No sooner had we started to celebrate than a fleet of ships surrounded our stolen colony vessel. Gaz took me by the arm when she heard the news. “The Originals got a warning out to Sirius-B29.” “So this isn’t a welcome party?” The ships escorted us through the planet’s atmosphere, and after our first steps onto real ground, we were put in chains and imprisoned.---a combination of people and aliens coming off spaceship onto the planet Sirius-B29 are put in chains by another human group, oil painting 
+~
+Philip waited in the dark. The countdown had long since hit zero, the call of "Ready or not..." rang through the trees. Philip crouched in the hollow of a tree beneath a blanket of branches, trembling with the anticipation of being found. Then he trembled because he feared the game was over, they'd forgotten all about him, and he was alone in the woods.---a boy hiding in a tree, waiting to be found in a game of hide-and-seek, pencil sketch
+~
+"You need to impress her, man," Mark said. Thomas looked down at his gawky body. "I don't know if I'm equipped for that." Mark shook his head. "Only the crude impress with their bodies. You can work her mind. You should learn magic." So that's what Thomas did. At first, he had no idea where to begin. He started reading books on magic and body language. He learned how to memorize strings of numbers and star constellations. He played mind games with people and practiced being disarming and charming. He became obsessed with magic, and he was good at it.--a man practicing magic in front of a mirror, digital 3D art, high resolution
+~
+"Is it just me, or is my sourdough starter saying hello?" Nicole added a bread emoji to the end of the Instagram post. She’d carefully framed the shot of bubbles in the flour mixture that appeared to spell out “hi.” Her heart lightened a little. She’d been dumped by Alex the day before, and baking kept her from spending the day crying. She started a new sourdough loaf and spent the day in a haze of memory and home-brewed beer.---a woman taking a picture of her sourdough starter that appears to say "hi", black and white photography
+~
+He said ' Sharon, what are you doing? The ship isn't even fixed yet!' She replied 'Look, I tried to fix the ship, but it kept blowing up! So, I tried to do something creative with the parts. And I think that this is the result.' She pointed to the creation. Andrey looked at it. It was a strange, weird thing. It looked like a metal face, with two eyes, a nose, and a mouth. 'What is it?' asked Andrey.And Sharon said, 'I don't really know, but I think it's a security camera.'---A woman shows a man her creation made out of the parts of a ship that kept blowing up, it looks like a metal face with two eyes, a nose, and mouth.
