+++ inputs/sample_input.txt
@@ -1,0 +1,81 @@
+Marlie nearly turned around when Reed stepped out of the souped-up armored truck, pounding an extra-large energy drink. This was not her scene, but she couldn't trust anyone else to deploy her drone system in the middle of the storm. If she wanted to finish her PhD thesis sometime in the next decade, she'd have to ride shotgun with this adrenaline-junkie and his video camera.
+
+Reed was a total blast from the past. Not just his uniform, but he was a descendant of that species of boys who used to crowd her high-school hallways, who would pinch her butt when she wasn't looking and then run away, laughing. Now, she had to drive all the way to fucking Oklahoma with him.
+
+"So, Dr. Ziele, did you say?”
+
+“Not Dr., yet. We get this storm data though, and I’ll be closer.”
+
+"You're no fun."
+
+"I'm a scientist, not a trained monkey. I don't work for the circus. Here, do you know how to use one of these?"
+
+She tossed him a pair of goggles and a monitor with an air-pressure tracker.
+
+Reed winced when the leather gloves slapped against his palm. "What's this?"
+
+"Our eyes on the storm. You'll be in the truck with me. I won't let you shout at me from the cab of the truck, so you'll have to do it from the mobile unit. Your deafening voice is part of the equipment, too."
+
+Reed downed the rest of his drink, then carefully placed the empty can on the truck's hood. He tapped the can with his finger, but then he suddenly realized Marli was watching him. He quickly wrapped the gloves around his fingers, adjusted the goggles so they were hanging from his neck, and tugged on the hood of his sweatshirt.
+
+"Right. I'm going to put this on. Then we'll take off. And we'll be on air. We'll show the world how we capture these images."
+
+Once he started his truck, Reed was a different type of animal, laser-focused on the road ahead and intercepting the monster supercell that formed overhead. Rain drummed on the roof of the cabin and Marlie had a bad feeling about the whole adventure.
+
+"There is is!" Reed screamed, pointing ahead at a giant tornado.
+
+Marlie had seen footage, had studied storms her whole career, and now, for the first time, the reality of a twister was seriously affecting her. She pulled her knees to her chest, gripped her seat belt and squeezed her eyes shut.
+
+"Keep staring at the road, moron," she muttered.
+
+"Sorry," he said, still bouncing around.
+
+"It's all right."
+
+"This is it, Marlie. This is why I became a storm chaser. This is my moment! This is--"
+
+"Yes, yes. I get it. You're excited."
+
+Reed's video camera was running. Marlie glanced at the screen where a funnel was clearly visible, slapping the earth with great force.
+
+"Turn back now," she said. "Let's go. We're out of here."
+
+"No! No--"
+
+"I said turn back!" Marli shrieked.
+
+Reed slammed on the brakes and threw open the door. He ran out into the screaming wind with his video camera.
+
+Marlie hugged her equipment to her chest. This was her moment, too, if she could just summon to courage to get out of the car and face down heavy weather. She fought her seat belt, but the latch wouldn't release.
+
+The tornado was right in front of them, destroying everything: trees, telephone poles, and all signs of civilization. The funnel extended into a massive, gray cone as tornadic winds blasted debris all around them.
+
+"Reed!" Marli screamed.
+
+The storm chaser had already been thrown several yards away by the storm, and the funnel was building up there, ready to destroy everything.
+
+"Reed!"
+
+A loud crash shattered the night. In the middle of Reed's fall, something had hit him on the head and he was down, and the funnel was coming right up on him.
+
+"Holy shit! Reed!"
+
+Marlie's belt finally released and she scrambled out of the truck. She had to help him. Even though they were on opposite sides of the political aisle, they were both people, and this was a meteorological emergency. She had to do the right thing. She dropped to her knees next to him. Reed's eyes were closed and blood was running down his forehead. He wasn't moving.
+
+The funnel was so close now, she could almost reach out and touch it. She was trapped, trapped with Reed under the funnel cloud.
+
+"Reed, open your eyes. Don't make me be the last person to see you alive. No fucking way."
+
+The tornado was closing in, but Marlie wasn't going to abandon Reed. She was on her own, there was nothing she could do to save herself, and she was going to be completely torn apart by the storm.
+
+Marlie grabbed Reed's camera and aimed it at her face. This was it.
+
+"This is Marlie Ziele, an atmospheric scientist from Columbia University. I am trapped under a tornado in Oklahoma. I am scared, and there's nothing I can do. My colleague, Reed Walker, is hurt, lying next to me. He's not moving. I'm trying to keep my shit together. Say something, Reed! Say something! If it's the last thing I ever do, I'll tell people about this."
+
+The funnel was so close now that she could see the individual elements: Wind. Water. Earth. Fire.
+
+"I'll tell the world about your heroism, Reed. You're the greatest goddamn hero I've ever known."
+
+The tornado struck and Marli was swept away into the vortex, screaming and screaming and screaming, and then she couldn't feel anything at all.
+
+
