+++ run.sh
@@ -3,8 +3,8 @@ if [[ ! -v OPENAI_TOKEN ]]; then
 elif [[ -z "$OPENAI_TOKEN" ]]; then
     echo "OPENAI_TOKEN is set to empty string"
 else
-    CUDA_VISIBLE_DEVICES=1 python run.py -f "$1" -n 2 &
-    CUDA_VISIBLE_DEVICES=0 python run.py -f "$1" -n 2 && fg
+    CUDA_VISIBLE_DEVICES=1 python scripts/run.py "$@" &
+    CUDA_VISIBLE_DEVICES=0 python scripts/run.py "$@" && fg
 fi
